pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "hrp"
version = "0.1.0"
description = "Fast numpy-only Hierarchical Risk Parity implementation"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.24.0",
    "scipy>=1.10.0",
]

[project.optional-dependencies]
dev = [
    "pandas>=2.0.0",
    "pytest>=7.0.0",
    "PyPortfolioOpt>=1.5.0",
    "riskfolio-lib>=6.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["hrp*"]
